+++ lib/federails/configuration.rb
@@ -42,6 +42,10 @@ module Federails
     mattr_accessor :client_routes_path
     @@client_routes_path = :app
 
+    # Custom URL to use for remote-following requests
+    mattr_accessor :remote_follow_url
+    @@remote_follow_url = nil
+
     # Method to use for links to user profiles
     # @deprecated Set profile_url_method option on acts_as_federails_actor instead
     mattr_accessor :user_profile_url_method